Refactor modules and move tests
